backend/requirements.txt
Flask==3.0.2
markdown==3.5.2
fpdf2==2.7.8
waitress==3.0.0
mcp==1.2.0
fastmcp==0.4.1
